refactor: extract combo-box-light logic to reusable mixin
